+++ tests/test_config/test_blank_line_after_tag.py
@@ -135,6 +135,23 @@ test_data = [
         ({"blank_line_after_tag": "   include     ,endblock "}),
         id="test multiple",
     ),
+    pytest.param(
+        (
+            "{%- macro Foo() -%}\n"
+            "  <div>foo</div>\n"
+            "{%- endmacro -%}\n"
+            "<p>after macro</p>\n"
+        ),
+        (
+            "{%- macro Foo() -%}\n"
+            "    <div>foo</div>\n"
+            "{%- endmacro -%}\n"
+            "\n"
+            "<p>after macro</p>\n"
+        ),
+        ({"blank_line_after_tag": "endmacro"}),
+        id="nunjucks whitespace control dash - after",
+    ),
 ]